file files along
* make the validator a library, and move the binary in a directory
* nom /validators only contains validators
