+++ src/daemon.rs
@@ -1,4 +1,4 @@
-use crate::validators::{create_issues_from_reader, process, Response};
+use crate::validate::{create_issues_from_reader, process, Response};
 use actix_web::{get, post, web, web::Json, App, Error, HttpServer, Responder};
 use futures::{future::ok, Future, Stream};
 use std::env;
