+++ src/lib.rs
@@ -1,0 +1,11 @@
+pub mod daemon;
+pub mod issues;
+pub mod metadatas;
+pub mod validate;
+pub mod validators;
+
+pub use issues::{Issue, IssueType, RelatedObject, Severity};
+pub use validate::{validate, validate_and_metadata};
+
+#[macro_use]
+extern crate serde_derive;
