+++ src/metadatas.rs
@@ -1,4 +1,4 @@
-use crate::validators::issues::IssueType;
+use crate::issues::IssueType;
 use gtfs_structures;
 use itertools::Itertools;
